+++ credit-service/src/main/java/ru/creditbank/credit/operations/dto/rs/ErrorRsDto.java
@@ -3,5 +3,5 @@ package ru.creditbank.credit.operations.dto.rs;
 import lombok.Builder;
 
 @Builder
-public record ErrorRsDto(String error) {
+public record ErrorRsDto(String message) {
 }