+++ credit-service/src/main/java/ru/creditbank/credit/operations/exception/RestException.java
@@ -1,0 +1,15 @@
+package ru.creditbank.credit.operations.exception;
+
+import lombok.Getter;
+import org.springframework.http.HttpStatus;
+
+@Getter
+public abstract class RestException extends RuntimeException {
+
+    private final HttpStatus httpStatus;
+
+    public RestException(String msg, HttpStatus httpStatus) {
+        super(msg);
+        this.httpStatus = httpStatus;
+    }
+}